--- tests/mock_api/db.py
@@ -1,5 +1,0 @@
-from .models.jobs import Job
-from .models.program import Program
-
-FAKE_JOB_DB: dict[str, Job] = {}
-FAKE_PROGRAM_DB: dict[str, Program] = {}
